app/requirements.txt
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
azure-functions
azure-storage-blob
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
